requirements.txt
scikit_image==0.15.0
torchvision==0.4.1a0+d94043a
tensorboard==2.0.1
pandas==0.25.2
h5py==2.10.0
scipy==1.3.1
Pillow==7.2.0
scikit_learn==0.23.1
